# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Billing & Solar ERP API",
        "description": "Billing + Cloud Solar Panel Installation & Service Management API. Workflow: Lead → Survey → Quotation → Installation → Service & AMC. Roles: MD Super Admin, Branch Manager, Branch Admin, Sales Executive, Site Survey Engineer, Installation Manager/Technician, Service Team, Accounts. JWT Secured: use POST /api/v1/auth/login then header Authorization: Bearer {access_token}. AI-friendly: structured JSON responses, clear field names, dashboard metrics.",
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server (localhost)"
        },
        {
            "url": "https://testbillapi.eazycutz.com/api/v1",
            "description": "API Server (Live)"
        },
        {
            "url": "/api/v1",
            "description": "API Server (relative)"
        }
    ],
    "paths": {
...
        "/solar/service-tickets": {
            "get": {
                "tags": [
                    "Solar - Service & AMC"
                ],
                "summary": "List Service Tickets",
                "description": "List service tickets.",
                "operationId": "469ed32a478d9d0a4d17250c1ba870bb",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "open",
                                "assigned",
                                "in_progress",
                                "resolved",
                                "closed"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Solar - Service & AMC"
                ],
                "summary": "Create Service Ticket",
                "description": "Create service ticket (Customer complaint / AMC).",
                "operationId": "8ef151a2b1ac4055e5328732ec6f78a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "complaint"
                                ],
                                "properties": {
                                    "installation_id": {
                                        "type": "integer"
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "complaint": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solar/service-tickets/{id}": {
            "get": {
                "tags": [
                    "Solar - Service & AMC"
                ],
                "summary": "Get Service Ticket",
                "description": "Get service ticket.",
                "operationId": "ea923e2343042c1e9be1f4240f981dda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Solar - Service & AMC"
                ],
                "summary": "Update Service Ticket",
                "description": "Update ticket (assign, resolve).",
                "operationId": "634320511126f59bf9fcc163e9ae031d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "assigned_to": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "assigned",
                                            "in_progress",
                                            "resolved",
                                            "closed"
                                        ]
                                    },
                                    "resolution": {
                                        "type": "string"
                                    },
                                    "scheduled_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: \\\"Authorization: Bearer {token}\\\"",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}